# api/openapi.yaml
openapi: 3.0.3
info:
  title: Dynamic Pricing (educational)
  version: 0.1.0
paths:
...
  /prices/{product_id}:
    servers:
      - url: http://localhost:8083
    get:
      tags: [Pricing]
      summary: Get current price
      parameters:
        - in: path
          name: product_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object